Check minifi dir
If there is no minifi dir, do nothing.

+++ scripts/restart_minifi.sh
@@ -5,6 +5,10 @@ MINIFI_HOME="$MINIFI_DIR/minifi-0.5.0"
 MINIFI_CONF="$MINIFI_HOME/conf"
 MINIFI_BIN="$MINIFI_HOME/bin"
 
-$MINIFI_BIN/minifi.sh stop
-rm -rf $MINIFI_HOME/content_repository/* $MINIFI_HOME/provenance_repository/* $MINIFI_HOME/flowfile_repository/* $MINIFI_HOME/state/local/* $MINIFI_HOME/logs/*
-$MINIFI_BIN/minifi.sh start
+if [ -d "$MINIFI_DIR" ]; then
+	$MINIFI_BIN/minifi.sh stop
+	rm -rf $MINIFI_HOME/content_repository/* $MINIFI_HOME/provenance_repository/* $MINIFI_HOME/flowfile_repository/* $MINIFI_HOME/state/local/* $MINIFI_HOME/logs/*
+	$MINIFI_BIN/minifi.sh start
+else
+	echo "$0: $MINIFI_DIR is missing."
+fi
