+++ scripts/start_minifi.sh
@@ -5,5 +5,9 @@ MINIFI_HOME="$MINIFI_DIR/minifi-0.5.0"
 MINIFI_CONF="$MINIFI_HOME/conf"
 MINIFI_BIN="$MINIFI_HOME/bin"
 
-rm -rf $MINIFI_HOME/content_repository/* $MINIFI_HOME/provenance_repository/* $MINIFI_HOME/flowfile_repository/* $MINIFI_HOME/state/local/* $MINIFI_HOME/logs/*
-$MINIFI_BIN/minifi.sh start
+if [ -d "$MINIFI_DIR" ]; then
+	rm -rf $MINIFI_HOME/content_repository/* $MINIFI_HOME/provenance_repository/* $MINIFI_HOME/flowfile_repository/* $MINIFI_HOME/state/local/* $MINIFI_HOME/logs/*
+	$MINIFI_BIN/minifi.sh start
+else
+	echo "$0: $MINIFI_DIR is missing."
+fi
